float percentages

+++ db/migrate/20150816023001_add_new_columns.rb
@@ -1,9 +1,9 @@
 class AddNewColumns < ActiveRecord::Migration
   def change
-    add_column :usages, :toilet, :integer, null: false
-    add_column :usages, :shower, :integer, null: false
-    add_column :usages, :laundry, :integer, null: false
-    add_column :usages, :kitchen, :integer, null: false
-    add_column :usages, :other, :integer, null: false
+    add_column :usages, :toilet, :float, null: false
+    add_column :usages, :shower, :float, null: false
+    add_column :usages, :laundry, :float, null: false
+    add_column :usages, :kitchen, :float, null: false
+    add_column :usages, :other, :float, null: false
   end
 end
